Moved CourseInfo into Webapi/output so that it gets printed, refactored inner classes CourseResponse and CoursesResponse to CourseData and CoursesData respectively and moved them to 2 separate files
