+++ src/main/java/teammates/ui/webapi/output/CoursesData.java
@@ -1,0 +1,21 @@
+package teammates.ui.webapi.output;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+import teammates.common.datatransfer.attributes.CourseAttributes;
+
+/**
+ * Response of a list of courses.
+ */
+public class CoursesData extends ApiOutput {
+    private final List<CourseData> courses;
+
+    public CoursesData(List<CourseAttributes> courseAttributesList) {
+        courses = courseAttributesList.stream().map(CourseData::new).collect(Collectors.toList());
+    }
+
+    public List<CourseData> getCourses() {
+        return courses;
+    }
+}
